diskusage: force types to be uint64
On freebsd it seems like Bavail is int64 but it will never be <0 so
let's just convert them into uint64 before multiplying to avoid issues.

+++ pkg/diskusage/diskusage_posix.go
@@ -18,8 +18,8 @@ func Get(path string) (Usage, error) {
 		return Usage{}, errors.Wrapf(err, "statfs %s", path)
 	}
 	return Usage{
-		FreeBytes:  stat.Bfree * uint64(stat.Bsize),
-		TotalBytes: stat.Blocks * uint64(stat.Bsize),
-		AvailBytes: stat.Bavail * uint64(stat.Bsize),
+		FreeBytes:  uint64(stat.Bfree) * uint64(stat.Bsize),
+		TotalBytes: uint64(stat.Blocks) * uint64(stat.Bsize),
+		AvailBytes: uint64(stat.Bavail) * uint64(stat.Bsize),
 	}, nil
 }
